Refactor sceneGraph structure
